Fixed data type mismatch

+++ domain-core/src/main/kotlin/tech/testra/reportal/domain/valueobjects/ValueObjects.kt
@@ -49,5 +49,5 @@ data class SimulationScenario(
 
 data class Percentile(
     val n: Byte,
-    val value: Long
+    val value: BigDecimal
 )
